add first version of status bar for import
