+++ RechteDB/templates/base.html
@@ -7,6 +7,7 @@
  		<link rel="stylesheet" href="{% static 'css/bootstrap.min.css' %}">
 		<link rel="stylesheet" href="{% static 'css/style.css' %}">
 		<link rel="shortcut icon" type="image/png" href="{% static 'favicon.ico' %}" />
+		<script src="{% static 'js/jquery-3.3.1.min.js' %}"></script>
 	</head>
 	<body>
 		<nav class="navbar sticky-top navbar-expand-lg navbar-expand-xl navbar-dark bg-primary">
